# reuse list removals for insertions in policy lists