old:
```
BenchmarkCache_memory-8   	       3	2070894136 ns/op	171105568 B/op	 3962128 allocs/op
PASS
ok  	github.com/graxinc/cache	12.127s
```

new:
```
BenchmarkCache_memory-8   	       3	1634825570 ns/op	115878714 B/op	 2837719 allocs/op
PASS
ok  	github.com/graxinc/cache	10.083s
```

bench cmp
```
BenchmarkCache_memory-8     2070894136     1634825570     -21.06%

benchmark                   old allocs     new allocs     delta
BenchmarkCache_memory-8     3962128        2837719        -28.38%

benchmark                   old bytes     new bytes     delta
BenchmarkCache_memory-8     171105568     115878714     -32.28%
```